refactor: add ThundercardApp class
