+++ lib/main.dart
@@ -1,15 +1,11 @@
-import 'package:dynamic_color/dynamic_color.dart';
+import 'package:firebase_app_check/firebase_app_check.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:firebase_app_check/firebase_app_check.dart';
-import 'package:thundercard/utils/create_theme.dart';
-
-import 'firebase_options.dart';
-import 'providers/custom_theme.dart';
-import 'views/pages/auth_gate.dart';
+import 'package:thundercard/app.dart';
+import 'package:thundercard/firebase_options.dart';
+import 'package:thundercard/providers/custom_theme.dart';
 
 final customThemeProvider = ChangeNotifierProvider((ref) {
   return CustomTheme();
@@ -17,7 +13,7 @@ final customThemeProvider = ChangeNotifierProvider((ref) {
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]);
+  await SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]);
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
@@ -26,42 +22,9 @@ void main() async {
         'recaptcha-v3-site-key', // If you're building a web app.
     // androidDebugProvider: true,
   );
-  runApp(const ProviderScope(child: MyApp()));
-}
-
-class MyApp extends ConsumerWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final customTheme = ref.watch(customThemeProvider);
-
-    return DynamicColorBuilder(
-      builder: (ColorScheme? lightDynamic, ColorScheme? darkDynamic) {
-        return MaterialApp(
-          debugShowCheckedModeBanner: false,
-          title: 'Thundercard',
-          theme: createTheme(lightDynamic, Brightness.light, context),
-          darkTheme: createTheme(darkDynamic, Brightness.dark, context),
-          themeMode: customTheme.currentAppTheme,
-          locale: const Locale('ja', 'JP'),
-          localizationsDelegates: const [
-            GlobalMaterialLocalizations.delegate,
-            GlobalWidgetsLocalizations.delegate,
-            GlobalCupertinoLocalizations.delegate,
-          ],
-          supportedLocales: const [
-            Locale('ja', 'JP'),
-          ],
-          home: AnnotatedRegion<SystemUiOverlayStyle>(
-            value: const SystemUiOverlayStyle(
-              statusBarColor: Colors.transparent,
-              systemNavigationBarColor: Colors.transparent,
-            ),
-            child: AuthGate(),
-          ),
-        );
-      },
-    );
-  }
+  runApp(
+    const ProviderScope(
+      child: ThundercardApp(),
+    ),
+  );
 }
